// src/api/services/swapi/models/people/IPeople.ts
export interface IPeople {
  birth_year: string; // The birth year of the person, using the in-universe standard of BBY or ABY - Before the Battle of Yavin or After the Battle of Yavin. The Battle of Yavin is a battle that occurs at the end of Star Wars episode IV: A New Hope.
  created: string; // the ISO 8601 date format of the time that this resource was created.
  edited: string; // the ISO 8601 date format of the time that this resource was edited.
  eye_color: string; // The eye color of this person. Will be "unknown" if not known or "n/a" if the person does not have an eye.
  films: Array<string>; // An array of film resource URLs that this person has been in.
  gender: string; // The gender of this person. Either "Male", "Female" or "unknown", "n/a" if the person does not have a gender.
  hair_color: string; // The hair color of this person. Will be "unknown" if not known or "n/a" if the person does not have hair.
  height: string; // The height of the person in centimeters.
  homeworld: string; // The URL of a planet resource, a planet that this person was born on or inhabits.
  mass: string; // The mass of the person in kilograms.
  name: string; // The name of this person.
  skin_color: string; // The skin color of this person.
  species: Array<string>; // An array of species resource URLs that this person belongs to.
  starships: Array<string>; // An array of starship resource URLs that this person has piloted.
  vehicles: Array<string>; // An array of vehicle resource URLs that this person has piloted.
  url: string; // the hypermedia URL of this resource.
}